Animations Algorithm
This is the plan so far. The 3 items that I have chosen are Counter, Puck Side, and Ever Larger, until not. I have also placed the directions that I need to follow for each part of the assignment. I intend on featuring all of these choices together on the same file, but I can't promise anything.