Objects Exercises 1 Complete!
Everything is as it should be. I greatly appreciate the help that I have been getting with these projects, and I feel like cooperation is something that really works in this industry. Of course, the display, tennis, and ball are as they should be, and hopefully there's no bugs in the system.